requirements.txt
pandas
numpy
scikit-learn
scipy
pandas_datareader
yfinance
streamlit
plotly
matplotlib
stqdm
st-annotated-text